% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBreathspec.R
\name{readBreathspec}
\alias{readBreathspec}
\title{Load breathspec data into a dataframe. Accepts either a vector of file
location, or a single file in which case every .csv is loaded.}
\usage{
readBreathspec(path)
}
\arguments{
\item{path}{Either a single folder, or a character vector of .csv files}
}
\value{
A dataframe containing breathspec data
}
\description{
Load breathspec data into a dataframe. Accepts either a vector of file
location, or a single file in which case every .csv is loaded.
}
\examples{
\dontrun{
df1 <- readBreathspec(c("breathspec1.csv", "breathspec2.csv"))
df2 <- readBreathspec("path/to/folder")
}
}
