% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBreathspec.R
\name{readBreathspecSingle}
\alias{readBreathspecSingle}
\title{Load a single breathspec file into a 1-row dataframe}
\usage{
readBreathspecSingle(file)
}
\arguments{
\item{A}{single .csv file}
}
\value{
Single-row dataframe
}
\description{
Load a single breathspec file into a 1-row dataframe
}
